+++ doc/assets/js/doc.js
@@ -4,11 +4,11 @@ var EqUIDoc = {};
 
     // Init
     EqUIDoc.site.init = function() {
-        
+
         // Set active path in side bar menu
-        if(_doc_route.root_path !== "" && _doc_route.path !== ""){
-            var _root_path = $('#collapsible-nav-'+_doc_route.root_path);
-            var _root_path_path = $('#collapsible-nav-'+_doc_route.root_path+'-'+_doc_route.path);
+        if (_doc_route.root_path !== "" && _doc_route.path !== "") {
+            var _root_path      = $('#collapsible-nav-' + _doc_route.root_path);
+            var _root_path_path = $('#collapsible-nav-' + _doc_route.root_path + '-' + _doc_route.path);
 
             // Set active for links
             _root_path.addClass("active");
@@ -17,8 +17,8 @@ var EqUIDoc = {};
             // Set for open collapsible item
             _root_path.parent().addClass("active");
 
-        } else if(_doc_route.path !== "") {
-            var _path = $('#collapsible-nav-'+_doc_route.path);
+        } else if (_doc_route.path !== "") {
+            var _path = $('#collapsible-nav-' + _doc_route.path);
 
             // Set active for links
             _path.addClass("active");
@@ -43,13 +43,13 @@ var EqUIDoc = {};
         }*/
 
         // Set Breadcrumb
-        var _breadcrumb = $('#main-eq-ui-app-bar').find('.eq-ui-breadcrumb');
-        var _breadcrumb_path = _doc_route.path === 'index' ? './':_doc_route.path+'.html';
-        if(_doc_route.name_section !== "" && _doc_route.name_chapter !== ""){
-            _breadcrumb.html('<a href="./" class="eq-ui-breadcrumb-item"><span>'+_doc_route.name_section+'</span></a>' +
-            '<a href="'+_breadcrumb_path+'" class="eq-ui-breadcrumb-item"><span>'+_doc_route.name_chapter+'</span></a>');
-        } else if(_doc_route.name_chapter !== "") {
-            _breadcrumb.html('<a href="'+_breadcrumb_path+'" class="eq-ui-breadcrumb-item"><span>'+_doc_route.name_chapter+'</span></a>');
+        var _breadcrumb      = $('#main-eq-ui-app-bar').find('.eq-ui-breadcrumb');
+        var _breadcrumb_path = _doc_route.path === 'index' ? './' : _doc_route.path + '.html';
+        if (_doc_route.name_section !== "" && _doc_route.name_chapter !== "") {
+            _breadcrumb.html('<a href="./" class="eq-ui-breadcrumb-item"><span>' + _doc_route.name_section + '</span></a>' +
+            '<a href="' + _breadcrumb_path + '" class="eq-ui-breadcrumb-item"><span>' + _doc_route.name_chapter + '</span></a>');
+        } else if (_doc_route.name_chapter !== "") {
+            _breadcrumb.html('<a href="' + _breadcrumb_path + '" class="eq-ui-breadcrumb-item"><span>' + _doc_route.name_chapter + '</span></a>');
         }
 
         // Refresh all collapsibles
@@ -57,51 +57,55 @@ var EqUIDoc = {};
 
         // Init modals
         $('.eq-ui-modal-trigger').leanModal({
-            dismissible: true, // Modal can be dismissed by clicking outside of the modal
-            opacity: .5, // Opacity of modal background
-            in_duration: 300, // Transition in duration
+            dismissible : true, // Modal can be dismissed by clicking outside of the modal
+            opacity     : .5, // Opacity of modal background
+            in_duration : 300, // Transition in duration
             out_duration: 200, // Transition out duration
-            ready: function() { console.log('Modal Open'); }, // Callback for Modal open
-            complete: function() { console.log('Modal Close'); } // Callback for Modal close
+            ready       : function () {
+                console.log('Modal Open');
+            }, // Callback for Modal open
+            complete    : function () {
+                console.log('Modal Close');
+            } // Callback for Modal close
         });
 
         // Form validations
-        if(_doc_route.path === "form-validation"){
+        if (_doc_route.path === "form-validation") {
             EqUIDoc.site.form_validations();
         }
 
         // Create File Upload Drop
         var _dropzone = EqUI.forms.file_upload_drop('.eq-ui-file-upload-drop-zone', {
-            url: 'http://localhost:8888/upload/',
+            url                  : 'http://localhost:8888/upload/',
             defaultImageThumbnail: "assets/img/doc/files/empty.png"
         });
 
-        if(_dropzone){
+        if (_dropzone) {
             // File Upload Drop - Events
-            _dropzone.on("addedfile", function(file) {
+            _dropzone.on("addedfile", function (file) {
                 console.log(file);
             });
         }
 
         // Build Git in Home
-        if(_doc_route.path === "index"){
+        if (_doc_route.path === "index") {
             EqUIDoc.site.build_git_home();
         }
 
         // Show layout structures
-        if(_doc_route.path === "buttons"){
+        if (_doc_route.path === "buttons") {
             EqUI.site.body.addClass('doc-show-in-layout-struture-1');
         }
 
         // Autocomplete
-        var autocomplete_minChars = 2;
+        var autocomplete_minChars  = 2;
         var autocomplete_countries = EqUIDoc.site.countries_v_d();
 
         // Autocomplete -> Top search
         $('#doc-eq-ui-top-search-autocomplete input').autocomplete({
-            lookup: autocomplete_countries,
+            lookup  : autocomplete_countries,
             appendTo: '#doc-eq-ui-top-search-autocomplete',
-            groupBy: '',
+            groupBy : '',
             minChars: autocomplete_minChars,
             onSelect: function (suggestion) {
                 alert('You selected: ' + suggestion.value + ', ' + suggestion.data);
@@ -110,9 +114,9 @@ var EqUIDoc = {};
 
         // Autocomplete -> Side nav -> search
         $('#doc-eq-ui-side-nav-search-autocomplete input').autocomplete({
-            lookup: autocomplete_countries,
+            lookup  : autocomplete_countries,
             appendTo: '#doc-eq-ui-side-nav-search-autocomplete',
-            groupBy: '',
+            groupBy : '',
             minChars: autocomplete_minChars,
             onSelect: function (suggestion) {
                 alert('You selected: ' + suggestion.value + ', ' + suggestion.data);
@@ -121,9 +125,9 @@ var EqUIDoc = {};
 
         // Autocomplete -> search
         $('#doc-eq-ui-search-autocomplete input').autocomplete({
-            lookup: autocomplete_countries,
+            lookup  : autocomplete_countries,
             appendTo: '#doc-eq-ui-search-autocomplete',
-            groupBy: '',
+            groupBy : '',
             minChars: autocomplete_minChars,
             onSelect: function (suggestion) {
                 alert('You selected: ' + suggestion.value + ', ' + suggestion.data);
@@ -132,9 +136,9 @@ var EqUIDoc = {};
 
         // Autocomplete -> search expandable
         $('#doc-eq-ui-search-expandable-autocomplete input').autocomplete({
-            lookup: autocomplete_countries,
+            lookup  : autocomplete_countries,
             appendTo: '#doc-eq-ui-search-expandable-autocomplete',
-            groupBy: '',
+            groupBy : '',
             minChars: autocomplete_minChars,
             onSelect: function (suggestion) {
                 alert('You selected: ' + suggestion.value + ', ' + suggestion.data);
@@ -143,17 +147,164 @@ var EqUIDoc = {};
 
         // Autocomplete -> input
         $('#doc-text-autocomplete input').autocomplete({
-            lookup: autocomplete_countries,
+            lookup  : autocomplete_countries,
             appendTo: '#doc-text-autocomplete',
-            groupBy: '',
+            groupBy : '',
             minChars: autocomplete_minChars,
             onSelect: function (suggestion) {
                 alert('You selected: ' + suggestion.value + ', ' + suggestion.data);
             }
         });
 
+        //-----------------------------------
+        // Table -> Lazy load (Sample 1)
+        //-----------------------------------
+        var fakeJsonData                           = "http://37.187.137.141:3500/posts"; // http://192.168.1.20:3500/posts
+        var fakeDataTable                          = [];
+        var EqUILazyLoad_Sample1                   = null;
+        var EqUILazyLoad_Sample1_Scroll_Content_ID = $('#table-lazy-load-sample1 tbody')[0];
+
+        // Load json data
+        if (EqUILazyLoad_Sample1_Scroll_Content_ID) {
+            $.getJSON(fakeJsonData, {
+                _page : 1,
+                _limit: 300000
+            })
+            .done(function (data) {
+
+                // Transform data
+                $.each(data, function (i, item) {
+                    fakeDataTable.push(
+                    '<tr>\n' +
+                    '<td class="eq-ui-data-table-cell-non-numeric table-fixed-col-100 eq-ui-hidden-xs-">' + (i + 1) + '</td>\n' +
+                    '<td class="eq-ui-data-table-cell-non-numeric"><span class="truncate">' + item.model + '</span></td>\n' +
+                    '<td>' + item.quantity + '</td>\n' +
+                    '<td>' + item.price + '</td>\n' +
+                    '</tr>'
+                    );
+                });
+
+                // Run Lazy Load
+                if (EqUILazyLoad_Sample1 === null) {
+                    EqUILazyLoad_Sample1 = new EqUILazyLoad({
+                        rows     : fakeDataTable,
+                        scrollId : EqUILazyLoad_Sample1_Scroll_Content_ID,
+                        contentId: EqUILazyLoad_Sample1_Scroll_Content_ID,
+                        callbacks: {
+                            clusterWillChange: function () {
+                                console.log('[Table -> Lazy load] clusterWillChange');
+                            },
+                            clusterChanged   : function () {
+                                console.log('[Table -> Lazy load] clusterChanged');
+                            },
+                            scrollingProgress: function (progress) {
+                                // console.log('[Table -> Lazy load] scrollingProgress', progress);
+                            }
+                        }
+                    });
+                }
+            });
+        }
+        //-----------------------------------
+
+        // Sticky Table
+        $(".eq-ui-sticky-table").EqUIStickyTable({top:50});
+
         // When our page loads, check to see if it contains and anchor
         EqUIDoc.site.scroll_if_anchor(window.location.hash);
+
+      //-----------------------------------
+      // Wassap Simulate
+      //-----------------------------------
+      /* Meme */
+      var memes = [
+        'Dude, you smashed my turtle saying "I\'M MARIO BROS!"',
+        'Dude, you grabed seven oranges and yelled "I GOT THE DRAGON BALLS!"',
+        'Dude, you threw my hamster across the room and said "PIKACHU I CHOOSE YOU!"',
+        'Dude, you congratulated a potato for getting a part in Toy Story',
+        'Dude, you were hugging an old man with a beard screaming "DUMBLEDORE YOU\'RE ALIVE!"',
+        'Dude, you were cutting all my pinapples yelling "SPONGEBOB! I KNOW YOU\'RE THERE!"',
+      ];
+
+      var random = document.querySelector('#random');
+
+      random.innerHTML = memes[Math.floor(Math.random() * memes.length)];
+
+      /* Time */
+      var deviceTime = document.querySelector('.status-bar .time');
+      var messageTime = document.querySelectorAll('.message .time');
+
+      deviceTime.innerHTML = moment().format('h:mm');
+
+      setInterval(function() {
+        deviceTime.innerHTML = moment().format('h:mm');
+      }, 1000);
+
+      for (var i = 0; i < messageTime.length; i++) {
+        messageTime[i].innerHTML = moment().format('h:mm A');
+      }
+
+      /* Message */
+      var form = document.querySelector('.conversation-compose');
+      var conversation = document.querySelector('.conversation-container');
+
+      form.addEventListener('submit', newMessage);
+
+      function newMessage(e) {
+        var input = e.target.input;
+        var guid = EqUI.guid();
+
+        if (input.value) {
+          var message = buildMessage(input.value, guid);
+          conversation.appendChild(message);
+          $('.dropdown-trigger-'+guid).dropdown();
+          animateMessage(message);
+        }
+
+        input.value = '';
+        conversation.scrollTop = conversation.scrollHeight;
+
+        e.preventDefault();
+      }
+
+      function buildMessage(text, guid) {
+        var element = document.createElement('div');
+
+        element.classList.add('message', 'sent');
+
+        element.innerHTML = text +
+          '<span class="metadata">' +
+          '<span class="time">' + moment().format('h:mm A') + '</span>' +
+          '<span class="tick tick-animation">' +
+          '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="15" id="msg-dblcheck" x="2047" y="2061"><path d="M15.01 3.316l-.478-.372a.365.365 0 0 0-.51.063L8.666 9.88a.32.32 0 0 1-.484.032l-.358-.325a.32.32 0 0 0-.484.032l-.378.48a.418.418 0 0 0 .036.54l1.32 1.267a.32.32 0 0 0 .484-.034l6.272-8.048a.366.366 0 0 0-.064-.512zm-4.1 0l-.478-.372a.365.365 0 0 0-.51.063L4.566 9.88a.32.32 0 0 1-.484.032L1.892 7.77a.366.366 0 0 0-.516.005l-.423.433a.364.364 0 0 0 .006.514l3.255 3.185a.32.32 0 0 0 .484-.033l6.272-8.048a.365.365 0 0 0-.063-.51z" fill="#92a58c"/></svg>' +
+          '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="15" id="msg-dblcheck-ack" x="2063" y="2076"><path d="M15.01 3.316l-.478-.372a.365.365 0 0 0-.51.063L8.666 9.88a.32.32 0 0 1-.484.032l-.358-.325a.32.32 0 0 0-.484.032l-.378.48a.418.418 0 0 0 .036.54l1.32 1.267a.32.32 0 0 0 .484-.034l6.272-8.048a.366.366 0 0 0-.064-.512zm-4.1 0l-.478-.372a.365.365 0 0 0-.51.063L4.566 9.88a.32.32 0 0 1-.484.032L1.892 7.77a.366.366 0 0 0-.516.005l-.423.433a.364.364 0 0 0 .006.514l3.255 3.185a.32.32 0 0 0 .484-.033l6.272-8.048a.365.365 0 0 0-.063-.51z" fill="#4fc3f7"/></svg>' +
+          '</span>' +
+          '</span>' +
+          '<div class="message-menu-option">' +
+          '<!-- Dropdown Trigger -->' +
+          '<a data-target="dropdown-message-menu-option-'+guid+'" data-auto-align-target="doc-conversation-container" data-hover="false" data-below-origin="true" class="dropdown-trigger-'+guid+'">' +
+          '<i class="mdi mdi-chevron-down icon icon-18"></i>' +
+          '</a>' +
+          '<!-- Dropdown Structure -->' +
+          '<ul id="dropdown-message-menu-option-'+guid+'" class="eq-ui-dropdown">' +
+          '<li><a href="#!">Some Action</a></li>' +
+          '<li class="disabled"><a href="#!">Disabled Action</a></li>' +
+          '<li><a href="#!" class="active">Item Active</a></li>' +
+          '<li><a href="#!">Another Action</a></li>' +
+          '</ul>' +
+          '</div>';
+
+        return element;
+      }
+
+      function animateMessage(message) {
+        setTimeout(function() {
+          var tick = message.querySelector('.tick');
+          tick.classList.remove('tick-animation');
+        }, 500);
+      }
+      //-----------------------------------
+
     };
 
     // Form validations