+++ src/js/buttons.js
@@ -68,13 +68,13 @@
 
         // End animation event
         function animationEnd() {
-            if(!element.hasClass(EqUI.buttons.fab_action_id+' active')){
+            if(!element.hasClass('active')){
                 $('.'+EqUI.buttons.fab_action_id+' ul').css('height', '0');
             }
         }
 
         // Add animation class
-        if(!element.hasClass(EqUI.buttons.fab_action_id+' active') && (isClose === false || isClose === 'none')){
+        if(!element.hasClass('active') && (isClose === false || isClose === 'none')){
             // SHOW
             $('.'+EqUI.buttons.fab_action_id+' ul').css('height', 'auto');
             element.addClass("active");
@@ -119,6 +119,18 @@
         }
     };
 
+
+    //Tooltip FAB
+
+
+    $( "a.eq-ui-btn-fab-with-tooltip" ).hover(
+        function() {
+            $( this ).siblings().addClass( "view-tooltip" );
+        }, function() {
+            $( this ).siblings().removeClass( "view-tooltip" );
+        }
+    );
+
     $(document).ready(function() {
         // Init
         EqUI.buttons.init();
@@ -126,4 +138,6 @@
         // Update
         EqUI.buttons.update();
     });
+
+
 }( jQuery ));