+++ src/js/layout/breadcrumb.js
@@ -83,7 +83,7 @@
                 var _last_children;
                 var _last_children_hiden;
                 if(origin_children.length === 1 && origin_children_hiden.length > 0){
-                    _last_children = $(origin_children[0]);
+                    // _last_children = $(origin_children[0]);
                     _last_children_hiden = $(origin_children_hiden[origin_children_hiden.length-1]);
                     origin.addClass(_this.breadcrumb_min_class);
                     // Save old last children hiden width