+++ src/js/tabs.js
@@ -47,6 +47,7 @@
                 if ($this.is(":visible")) {
                     $indicator.css({"right": $tab_right});
                     $indicator.css({"left": $tab_left});
+                    $indicator.css({"width": $tab_width});
                 }
 
                 // Resize
@@ -64,6 +65,7 @@
                     if ($tab_width !== 0 && $tabs_width !== 0) {
                         $indicator.css({"right": $tab_right});
                         $indicator.css({"left": $tab_left});
+                        $indicator.css({"width": $tab_width});
                     }
                 });
 
@@ -114,10 +116,12 @@
                     if (($index - $prev_index) >= 0) {
                         $indicator.velocity({"right": $tab_right}, { duration: 300, queue: false, easing: 'easeOutQuad'});
                         $indicator.velocity({"left": $tab_left}, {duration: 300, queue: false, easing: 'easeOutQuad', delay: 90});
+                        $indicator.velocity({"width": $tab_width});
                     }
                     else {
                         $indicator.velocity({"left": $tab_left}, { duration: 300, queue: false, easing: 'easeOutQuad'});
                         $indicator.velocity({"right": $tab_right}, {duration: 300, queue: false, easing: 'easeOutQuad', delay: 90});
+                        $indicator.velocity({"width": $tab_width});
                     }
 
                     // Prevent the anchor's default click action