+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 ## CHANGELOG
 
+### v0.4.0 (Nov 12, 2018)
+
+**Updates**
+- Add new [Auto Align - Dropdown](http://bootstrap-material-ui.exentriq.com/dropdown.html#auto-align)
+
+### v0.3.9 (Aug 27, 2018)
+
+**Updates**
+- Add new [Sticky Table](http://bootstrap-material-ui.exentriq.com/sticky-table.html)
+
+**Fixes**
+- Focus in top Search
+- Others fixes
+
+### v0.3.8 (Aug 10, 2018)
+
+**Updates**
+- Add new [Lazy Load For Table](http://bootstrap-material-ui.exentriq.com/table.html#table-lazy-load)
+
 ### v0.3.7 (Dec 18, 2017)
 
 **Updates**